Import Legacy application
Bharath provides a legacy application for this section.

The legacy application was cleaned up as follows:

1. Add `.gitignore`
1. Update dependencies support Java 15
1. Update `pom.xml`

The project as provided contains a `.DS_Store` for Mac OS. This is removed and ignored.

The `.gitignore` file is copied from the Spring Initializr template, edited as noted in previous
lessons.

Java 15 support is available in Apache CXF and the `pom.xml` is updated to reflect this support
with dependencies for Spring Boot and CXF releases.